requirements.txt
gradio
wikipedia
mcp
langchain
langchain-openai
langgraph
langchain-mcp-adapters
dotenv